代码收尾
代码收尾

+++ smart-report-backed/src/main/java/com/gsxy/core/pojo/User.java
@@ -18,7 +18,7 @@ public class User {
     //密码
     private String password;
     //权限
-    private Long role;
+    private Integer role;
     //创建时间
     private Date createdTime;
    // 修改人
